learning_users/requirements.txt
Django==5.1.7
django-phonenumber-field 8.0.0